mail-bucket/requirements.txt
dapr==1.15.0
fastapi==0.115.12
uvicorn==0.23.2
grpcio>=1.73.0
pydantic==2.4.2
requests==2.31.0
